fix compile errors in c files

+++ ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: jbonnin <jbonnin@42student.fr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/09 15:31:34 by jbonnin           #+#    #+#             */
-/*   Updated: 2023/01/14 19:07:39 by jbonnin          ###   ########.fr       */
+/*   Updated: 2023/01/14 20:10:51 by jbonnin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdio.h>
 
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, int n)
 {
 	int		i;
 	char	*res;
